+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fixed there being a `Nexus_Wallet` AND `Nexus Wallet` folders, Removed `Nexus_Wallet` and will auto move items into correct folder
 - Fixed issue with Recieved transactions not being shown in the table when selecting Receive Only Filter.
 - Fixed issue where the daemon would not be set to verbose 0, and set a min/max on the setting
+- Fixed issue where you could not copy the Address or Account from a transaction
 
 # 1.0.1 (2019.6.28)